requirements.txt
asgiref==3.5.2
boto3==1.24.47
botocore==1.27.47
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
dj-database-url==1.0.0
Django==3.2.14
django-ckeditor==6.4.2
django-cors-headers==3.13.0
django-debug-toolbar==3.5.0
django-js-asset==2.0.0
django-storages==1.13.1
django-taggit==3.0.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.20.0
Faker==12.0.1
google-api-core==2.8.2
google-auth==2.9.1
googleapis-common-protos==1.56.4
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
model-bakery==1.7.0
oauthlib==3.2.0
packaging==21.3
Pillow==9.2.0
protobuf==4.21.2
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
rest-framework-generic-relations==2.1.0
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.10
